Update patch set 1
Patch Set 1:

Not quite sure if we'd have a reason not to show these options,
especially knowing it's in a very specific menu for broadcast
and that only concerned users usually go into it.

I enable it for India & Brazil users who receive up to one broadcast every 5 minutes.

Patch-set: 1