Update patch set 4
Patch Set 4:

> (1 comment)

Please first fix phone app crashes all the time and also sim cards not recognizing on the latest builds 8, 9, 10, 11, 12.
This is major issue please try to fix this issue on priority. Only stable version is 7.

This issue is happening in Moto G3 2015.
Thanks for understanding.

Due to this we are unable to get features or issues resolved in latest builds.

Patch-set: 4