Update patch set 1
Patch Set 1:

Do we know for sure, if the automatic showing of these settings, based on SIM cards detected from those countries, is consistently working? I mean, it could work for some SIM cards, and not for others maybe?

I guess it can't hurt leaving it as it is.

Patch-set: 1