Update patch set 1
Patch Set 1: Code-Review+1

Tested using the four major Brazilian operators, all worked fine. Not sure about Indians.

Also, I agree it won't hurt, but I think we shouldn't display settings that will not work. But if we are not sure if Indians will have problems, the best option is to let them visible (although it is set to 'false' in AOSP code).

Patch-set: 1
Label: Code-Review=+1